iteration2 step 1

+++ src/main/java/ru/javacourse/eventmanagement/repository/LocationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.javacourse.eventmanagement.entity.location.LocationEntity;
 
 @Repository
-public interface LocationRepository extends JpaRepository<LocationEntity, Integer> {
+public interface LocationRepository extends JpaRepository<LocationEntity, Long> {
 
 }
