+++ src/main/java/ru/javacourse/eventmanagement/web/controllers/LocationController.java
@@ -38,14 +38,14 @@ public class LocationController {
     }
 
     @DeleteMapping("/{locationId}")
-    public ResponseEntity<LocationDto> deleteLocation(@PathVariable Integer locationId) {
+    public ResponseEntity<LocationDto> deleteLocation(@PathVariable Long locationId) {
         var location = locationService.deleteLocation(locationId);
         log.info("Delete a location {}", location);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(locationMapper.mapToDto(location));
     }
 
     @GetMapping("/{locationId}")
-    public ResponseEntity<LocationDto> getLocationById(@PathVariable Integer locationId) {
+    public ResponseEntity<LocationDto> getLocationById(@PathVariable Long locationId) {
         var locationByID = locationService.getLocationByID(locationId);
         log.info("Get a location {}", locationByID);
         return ResponseEntity.status(HttpStatus.OK).body(locationMapper.mapToDto(locationByID));
@@ -53,7 +53,7 @@ public class LocationController {
 
 
     @PutMapping("/{locationId}")
-    public ResponseEntity<LocationDto> updateLocationByID(@PathVariable  Integer locationId
+    public ResponseEntity<LocationDto> updateLocationByID(@PathVariable  Long locationId
             , @RequestBody LocationDto locationDto) {
         var eventLocation = locationMapper.mapFromDto(locationDto);
         Location locationUpdate = locationService.updateById(locationId,eventLocation);
